move checkbox styled component to ui package + separate types, wrappers
