+++ packages/renderer/src/renderer/screens/Action/DropdownMenus/types.ts
@@ -2,6 +2,13 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import type { TxStatus } from 'packages/types/src';
+import type { AnyData } from '@polkadot-live/types/misc';
+
+export interface CheckboxRxProps {
+  selected: boolean;
+  theme: AnyData;
+  onChecked: () => void;
+}
 
 export interface ExtrinsicDropdownMenuProps {
   isBuilt: boolean;
