+++ packages/ui/src/styles/index.ts
@@ -1,7 +1,9 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-export * from './ScrollWrappers';
-export * from './StatsFooter';
+export * from './checkbox';
 export * from './flex';
 export * from './grid';
+
+export * from './ScrollWrappers';
+export * from './StatsFooter';
